% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZarrColumnSeed.R
\name{ZarrColumnSeed}
\alias{ZarrColumnSeed}
\alias{ZarrColumnSeed-class}
\alias{dim,ZarrColumnSeed-method}
\alias{type,ZarrColumnSeed-method}
\alias{path,ZarrColumnSeed-method}
\alias{extract_array,ZarrColumnSeed-method}
\alias{ZarrColumnVector-class}
\alias{DelayedArray,ZarrColumnSeed-method}
\alias{ZarrColumnVector}
\title{Column of an Zarr-based data frame}
\usage{
ZarrColumnSeed(path, name, column, type = NULL, length = NULL)

ZarrColumnVector(x, ...)
}
\arguments{
\item{path}{String containing a path to a Zarr-based data frame.}

\item{name}{String containing the Zarr group of the Zarr file.}

\item{column}{String containing the name of the column inside the Zarr file.}

\item{type}{String specifying the type of the data.
If \code{NULL}, this is determined by inspecting the file.
Users may specify this to avoid a look-up, or to coerce the output into a different type.}

\item{length}{Integer containing the number of rows.
If \code{NULL}, this is determined by inspecting the file.
This should only be supplied for efficiency purposes, to avoid a file look-up on construction.}

\item{x}{Either a string containing the path to an Zarr-based data frame file (to be used as \code{path}),
or an existing ZarrColumnSeed object.}

\item{...}{Further arguments to be passed to the \code{ZarrColumnSeed} constructor.}
}
\value{
For \code{ZarrColumnSeed}, a ZarrColumnSeed is returned, obviously.

For \code{ZarrColumnVector}, a ZarrColumnVector is returned.
}
\description{
Represent a column of a Zarr-based data frame as a 1-dimensional \linkS4class{DelayedArray}.
This allows us to use Zarr-backed data inside \linkS4class{DataFrame}s without loading them into memory.
}
\author{
Artür Manukyan
}
