% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZarrDataFrame.R
\name{ZarrDataFrame}
\alias{ZarrDataFrame}
\alias{ZarrDataFrame-class}
\alias{nrow,ZarrDataFrame-method}
\alias{ncol,ZarrDataFrame-method}
\alias{length,ZarrDataFrame-method}
\alias{path,ZarrDataFrame-method}
\alias{rownames,ZarrDataFrame-method}
\alias{names,ZarrDataFrame-method}
\alias{rownames<-,ZarrDataFrame-method}
\alias{names<-,ZarrDataFrame-method}
\alias{extractROWS,ZarrDataFrame,ANY-method}
\alias{extractCOLS,ZarrDataFrame-method}
\alias{[[,ZarrDataFrame-method}
\alias{replaceROWS,ZarrDataFrame-method}
\alias{replaceCOLS,ZarrDataFrame-method}
\alias{normalizeSingleBracketReplacementValue,ZarrDataFrame-method}
\alias{[[<-,ZarrDataFrame-method}
\alias{cbind,ZarrDataFrame-method}
\alias{cbind.ZarrDataFrame}
\alias{as.data.frame,ZarrDataFrame-method}
\alias{coerce,ZarrDataFrame,DFrame-method}
\title{Zarr-backed DataFrame}
\usage{
ZarrDataFrame(tab, name, columns = NULL, nrows = NULL)
}
\arguments{
\item{tab}{A set of ZarrArrays that are the columns of a data frame.}

\item{name}{String containing the Zarr group of the Zarr file.}

\item{columns}{Character vector containing the names of columns in a  Zarr-based data frame.
If \code{NULL}, this is determined from \code{path}.}

\item{nrows}{Integer scalar specifying the number of rows in a  Zarr-based data frame.
If \code{NULL}, this is determined from \code{path}.}
}
\value{
A ZarrDataFrame where each column is a \linkS4class{ZarrColumnVector}.
}
\description{
Create a Zarr-backed \linkS4class{DataFrame}, where the data are kept on disk until requested.
}
\author{
Artür Manukyan
}
